project/requirements/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
Werkzeug==2.3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.15
WTForms==3.0.1
itsdangerous==2.1.2